fix(contract): corrige la suma i128 + u128 en el test de reclamo
token_client.balance devuelve i128 y claim_revenue devuelve u128, asi que la
asercion del test de reclamo no compilaba y la suite entera de la rama
feat/xlm-payments nunca llego a ejecutarse. Con el cast pasan los 13 tests.

+++ contracts/niko_project/src/lib.rs
@@ -1028,7 +1028,7 @@ mod test {
         let claimed = client.claim_revenue(&buyer, &project_id);
         assert!(claimed > 0);
         assert_eq!(claimed, claimable);
-        assert_eq!(token_client.balance(&buyer), buyer_before + claimed);
+        assert_eq!(token_client.balance(&buyer), buyer_before + claimed as i128);
     }
 
     /// Regression test: a buyer who purchases AFTER a revenue deposit has
